% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ete.R
\name{download_ete}
\alias{download_ete}
\title{Download the ETE data}
\usage{
download_ete(path = get_default_data_path(), version = "latest", quiet = FALSE)
}
\arguments{
\item{path}{Folder into which data will be downloaded}

\item{version}{Version of the data to download (default = "latest")}

\item{quiet}{logical, whether to download data silently}
}
\value{
None
}
\description{
This downloads the latest ETE data regardless if they are
  actually updated or not.
}
\examples{
\donttest{
  download_ete()
}

}
