% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ete.R
\name{get_data_versions}
\alias{get_data_versions}
\title{get version and download info for ETE data}
\usage{
get_data_versions()
}
\arguments{
\item{halt_on_error}{logical; if `FALSE`, return NULL on errors, otherwise
whatever got returned (could be an error or warning)}
}
\value{
A data.frame with two columns, `version` (string with the version #) and
  `zipball_url` (download URLs for the corresponding zipped release).
}
\description{
Check FigShare for the version and download link
  for ETE. This is fake for now, since we only have 1 version.
}
