% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ete_functions.R
\name{getages}
\alias{getages}
\title{Extract site ages in ybp}
\usage{
getages(data_tables, provider_name)
}
\arguments{
\item{data_tables}{Variable that holds the ete data tables}

\item{provider_name}{Name of the dataset provider}
}
\value{
A dataframe of sitekey and ages from a given provider
}
\description{
This extracts the site ages by provider.
}
\examples{
\donttest{
  ages <- getages(data_tables, "Amatangelo")
}

}
