% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ete_functions.R
\name{geteteoccur}
\alias{geteteoccur}
\title{Extract occurrence data by provider in long format}
\usage{
geteteoccur(data_tables, provider_name)
}
\arguments{
\item{data_tables}{Variable that holds the ete data tables}

\item{provider_name}{Name of the dataset provider}
}
\value{
A dataframe of occurrence data from a given provider
}
\description{
This extracts the occurrence data by provider.
}
\examples{
\donttest{
  amatangelo_occur <- geteteoccur(dataset_tables, 'Amatangelo')
}

}
