% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ete_functions.R
\name{geteteoccurDataset}
\alias{geteteoccurDataset}
\title{Extract occurrence data by dataset}
\usage{
geteteoccurDataset(data_tables, dataset_name)
}
\arguments{
\item{data_tables}{Variable that holds the ete data tables}

\item{pdataset_name}{Name of the dataset}
}
\value{
A dataframe of occurrence data from a given dataset
}
\description{
This extracts the occurrence data by dataset.
}
\examples{
\donttest{
  occur <- geteteoccurDataset(dataset_tables, 'Amatan_WI_Pla_Mod')
}

}
