% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ete_functions.R
\name{getlatlon}
\alias{getlatlon}
\title{Extract site coordinates by provider}
\usage{
getlatlon(data_tables, provider_name)
}
\arguments{
\item{data_tables}{Variable that holds the ete data tables}

\item{provider_name}{Name of the dataset provider}
}
\value{
A dataframe of site coordinates from a given provider
}
\description{
This extracts the site coordinates by provider.
}
\examples{
\donttest{
  wing_sites <- getlatlon('Wing')
}

}
