% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ete_functions.R
\name{getspptraits}
\alias{getspptraits}
\title{Extract sprecies trait matrix}
\usage{
getspptraits(data_tables, provider_name)
}
\arguments{
\item{data_tables}{Variable that holds the ete data tables}

\item{provider_name}{Name of the dataset provider}
}
\value{
A dataframe of species traits
}
\description{
This extracts a species trait matrix by provider.
}
\examples{
\donttest{
  spptraits <- getspptraits(data_tables,"Amatangelo")
}

}
