% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ete.R
\name{load_ete_data}
\alias{load_ete_data}
\title{The functions load_ete_data and load_datafile were both borrowed
heavily from https://github.com/weecology/portalr/blob/master/R/load_data.R,
with permission of the portalr maintainers.}
\usage{
load_ete_data(
  path = get_default_data_path(),
  download_if_missing = TRUE,
  clean = TRUE,
  quiet = FALSE
)
}
\value{
\code{\link{load_ete_data}} returns a list of 6 dataframes:
  \tabular{ll}{
    \code{dataset_table} \tab description\cr
    \code{occurrence_table} \tab description\cr
    \code{sites_table} \tab description\cr
    \code{sitetrait_table} \tab description\cr
    \code{species_table} \tab description\cr
    \code{speciestrait_table} \tab description\cr
  }
}
\description{
loads the ETE data files
}
\examples{
\donttest{
ete_data <- load_ete_data()
}
}
