% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ete_functions.R
\name{unmelt2specXsite}
\alias{unmelt2specXsite}
\title{Put your occurrence table in P/A matrix format}
\usage{
unmelt2specXsite(df)
}
\arguments{
\item{dataframe}{Dataframe to unmelt}
}
\value{
A dataframe of occurrence data in P/A matrix format
}
\description{
This melts your occurrence table in P/A matrix format
}
\examples{
\donttest{
  PAtable <- unmelt2specXsite(occurrences)
}

}
